src/requirements.txt
split-folders
numpy
torch
torchvision
pytorch_lightning
Pillow
tensorboard

 
